+++ src/app/api/user/userService.ts
@@ -1,0 +1,16 @@
+import { User } from '@/shared/interfaces'
+
+export class AdminService {
+
+  async create(user:User){
+    
+  }
+
+  async delete(id: number) {
+
+  }
+  
+  async getAll(){
+
+  }
+}
