+++ src/shared/interfaces/User.ts
@@ -1,4 +1,9 @@
 export interface User{
+  id: number,
+  fullname: string,
   username: string,
-  password: string
+  password: string,
+  createdAt: Date,
+  isActive: boolean,
+  roleId: number
 }
